backend/requirements.txt
fastapi==0.127.1
uvicorn==0.40.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.1
passlib==1.7.4
bcrypt==4.0.1
pydantic>=2.10.5
python-multipart==0.0.21
python-jose[cryptography]
fastapi-mail==1.6.2
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
alembic==1.18.4